refactor: 신규 사용자 주간 리포트 조회 로직 수정 #143
- 사용자의 daily_limit 값이 DB에 설정되어 있는지 확인
- 설정된 값이 없을 경우, BudgetService를 호출하여 '추천 한도'를 계산하고 그 값을 리포트 생성의 기준 한도로 사용
- 설정된 값이 있을 경우, 기존과 같이 해당 값을 기준 한도로 사용
