+++ src/main/java/cheongsan/domain/deposit/service/ReportCreationService.java
@@ -1,0 +1,7 @@
+package cheongsan.domain.deposit.service;
+
+import cheongsan.domain.deposit.dto.WeeklyReportDTO;
+
+public interface ReportCreationService {
+    WeeklyReportDTO createAndSaveLatestWeeklyReport(Long userId);
+}
